Severe-Artifact-CT-Segmentation/requirements.txt
conda
fvcore==0.1.5.post20221221
loralib==0.1.2
matplotlib==3.6.2
numpy==1.23.4
Pillow==9.2.0
Pillow==10.4.0
scikit_learn==1.5.1
tensorboardX==2.6.2.2
timm==1.0.9
torch==2.4.0
torchmetrics==1.4.1
torchvision==0.19.0
tqdm==4.66.1
xformers==0.0.27.post2
